+++ system/extras/simpleperf/record_file_writer.cpp
@@ -18,8 +18,8 @@
 
 #include <fcntl.h>
 #include <string.h>
-#include <sys/mman.h>
 #include <unistd.h>
+
 #include <algorithm>
 #include <set>
 #include <string>
@@ -33,7 +33,10 @@
 #include "event_attr.h"
 #include "perf_event.h"
 #include "record.h"
+#include "system/extras/simpleperf/record_file.pb.h"
 #include "utils.h"
+
+namespace simpleperf {
 
 using namespace PerfFileFormat;
 
@@ -50,22 +53,22 @@ std::unique_ptr<RecordFileWriter> RecordFileWriter::CreateInstance(const std::st
     return nullptr;
   }
 
-  return std::unique_ptr<RecordFileWriter>(new RecordFileWriter(filename, fp));
+  return std::unique_ptr<RecordFileWriter>(new RecordFileWriter(filename, fp, true));
 }
 
-RecordFileWriter::RecordFileWriter(const std::string& filename, FILE* fp)
+RecordFileWriter::RecordFileWriter(const std::string& filename, FILE* fp, bool own_fp)
     : filename_(filename),
       record_fp_(fp),
+      own_fp_(own_fp),
       attr_section_offset_(0),
       attr_section_size_(0),
       data_section_offset_(0),
       data_section_size_(0),
       feature_section_offset_(0),
-      feature_count_(0) {
-}
+      feature_count_(0) {}
 
 RecordFileWriter::~RecordFileWriter() {
-  if (record_fp_ != nullptr) {
+  if (record_fp_ != nullptr && own_fp_) {
     fclose(record_fp_);
     unlink(filename_.c_str());
   }
@@ -316,83 +319,88 @@ bool RecordFileWriter::WriteAuxTraceFeature(const std::vector<uint64_t>& auxtrac
     data.push_back(offset);
     data.push_back(AuxTraceRecord::Size());
   }
-  return WriteFeatureBegin(FEAT_AUXTRACE) && Write(data.data(), data.size() * sizeof(uint64_t)) &&
-         WriteFeatureEnd(FEAT_AUXTRACE);
+  return WriteFeature(FEAT_AUXTRACE, reinterpret_cast<char*>(data.data()),
+                      data.size() * sizeof(uint64_t));
 }
 
-bool RecordFileWriter::WriteFileFeatures(const std::vector<Dso*>& files) {
-  for (Dso* dso : files) {
+bool RecordFileWriter::WriteFileFeatures(const std::vector<Dso*>& dsos) {
+  for (Dso* dso : dsos) {
     // Always want to dump dex file offsets for DSO_DEX_FILE type.
     if (!dso->HasDumpId() && dso->type() != DSO_DEX_FILE) {
       continue;
     }
-    uint32_t dso_type = dso->type();
-    uint64_t min_vaddr;
-    uint64_t file_offset_of_min_vaddr;
-    dso->GetMinExecutableVaddr(&min_vaddr, &file_offset_of_min_vaddr);
+    FileFeature file;
+    file.path = dso->Path();
+    file.type = dso->type();
+    dso->GetMinExecutableVaddr(&file.min_vaddr, &file.file_offset_of_min_vaddr);
 
     // Dumping all symbols in hit files takes too much space, so only dump
     // needed symbols.
     const std::vector<Symbol>& symbols = dso->GetSymbols();
-    std::vector<const Symbol*> dump_symbols;
     for (const auto& sym : symbols) {
       if (sym.HasDumpId()) {
-        dump_symbols.push_back(&sym);
+        file.symbol_ptrs.emplace_back(&sym);
       }
     }
-    std::sort(dump_symbols.begin(), dump_symbols.end(), Symbol::CompareByAddr);
+    std::sort(file.symbol_ptrs.begin(), file.symbol_ptrs.end(), Symbol::CompareByAddr);
 
-    const std::vector<uint64_t>* dex_file_offsets = dso->DexFileOffsets();
-    if (!WriteFileFeature(dso->Path(), dso_type, min_vaddr, file_offset_of_min_vaddr,
-                          dump_symbols, dex_file_offsets)) {
+    if (const auto dex_file_offsets = dso->DexFileOffsets(); dex_file_offsets != nullptr) {
+      file.dex_file_offsets = *dex_file_offsets;
+    }
+    if (!WriteFileFeature(file)) {
       return false;
     }
   }
   return true;
 }
 
-bool RecordFileWriter::WriteFileFeature(const std::string& file_path,
-                                        uint32_t file_type,
-                                        uint64_t min_vaddr,
-                                        uint64_t file_offset_of_min_vaddr,
-                                        const std::vector<const Symbol*>& symbols,
-                                        const std::vector<uint64_t>* dex_file_offsets) {
-  uint32_t size = file_path.size() + 1 + sizeof(uint32_t) * 2 +
-      sizeof(uint64_t) + symbols.size() * (sizeof(uint64_t) + sizeof(uint32_t));
-  for (const auto& symbol : symbols) {
+bool RecordFileWriter::WriteFileFeature(const FileFeature& file) {
+  uint32_t symbol_count = file.symbols.size() + file.symbol_ptrs.size();
+  uint32_t size = file.path.size() + 1 + sizeof(uint32_t) * 2 + sizeof(uint64_t) +
+                  symbol_count * (sizeof(uint64_t) + sizeof(uint32_t));
+  for (const auto& symbol : file.symbols) {
+    size += strlen(symbol.Name()) + 1;
+  }
+  for (const auto& symbol : file.symbol_ptrs) {
     size += strlen(symbol->Name()) + 1;
   }
-  if (dex_file_offsets != nullptr) {
-    size += sizeof(uint32_t) + sizeof(uint64_t) * dex_file_offsets->size();
+  if (file.type == DSO_DEX_FILE) {
+    size += sizeof(uint32_t) + sizeof(uint64_t) * file.dex_file_offsets.size();
   }
-  if (file_type == DSO_ELF_FILE) {
+  if (file.type == DSO_ELF_FILE || file.type == DSO_KERNEL_MODULE) {
     size += sizeof(uint64_t);
   }
   std::vector<char> buf(sizeof(uint32_t) + size);
   char* p = buf.data();
   MoveToBinaryFormat(size, p);
-  MoveToBinaryFormat(file_path.c_str(), file_path.size() + 1, p);
-  MoveToBinaryFormat(file_type, p);
-  MoveToBinaryFormat(min_vaddr, p);
-  uint32_t symbol_count = static_cast<uint32_t>(symbols.size());
+  MoveToBinaryFormat(file.path.c_str(), file.path.size() + 1, p);
+  MoveToBinaryFormat(static_cast<uint32_t>(file.type), p);
+  MoveToBinaryFormat(file.min_vaddr, p);
   MoveToBinaryFormat(symbol_count, p);
-  for (const auto& symbol : symbols) {
+
+  auto write_symbol = [&](const Symbol* symbol) {
     MoveToBinaryFormat(symbol->addr, p);
     uint32_t len = symbol->len;
     MoveToBinaryFormat(len, p);
     MoveToBinaryFormat(symbol->Name(), strlen(symbol->Name()) + 1, p);
+  };
+  for (const auto& symbol : file.symbols) {
+    write_symbol(&symbol);
   }
-  if (dex_file_offsets != nullptr) {
-    uint32_t offset_count = dex_file_offsets->size();
+  for (const auto& symbol : file.symbol_ptrs) {
+    write_symbol(symbol);
+  }
+  if (file.type == DSO_DEX_FILE) {
+    uint32_t offset_count = file.dex_file_offsets.size();
     MoveToBinaryFormat(offset_count, p);
-    MoveToBinaryFormat(dex_file_offsets->data(), offset_count, p);
+    MoveToBinaryFormat(file.dex_file_offsets.data(), offset_count, p);
   }
-  if (file_type == DSO_ELF_FILE) {
-    MoveToBinaryFormat(file_offset_of_min_vaddr, p);
+  if (file.type == DSO_ELF_FILE || file.type == DSO_KERNEL_MODULE) {
+    MoveToBinaryFormat(file.file_offset_of_min_vaddr, p);
   }
   CHECK_EQ(buf.size(), static_cast<size_t>(p - buf.data()));
 
-  return WriteFeature(FEAT_FILE, buf);
+  return WriteFeature(FEAT_FILE, buf.data(), buf.size());
 }
 
 bool RecordFileWriter::WriteMetaInfoFeature(
@@ -408,11 +416,27 @@ bool RecordFileWriter::WriteMetaInfoFeature(
     MoveToBinaryFormat(pair.first.c_str(), pair.first.size() + 1, p);
     MoveToBinaryFormat(pair.second.c_str(), pair.second.size() + 1, p);
   }
-  return WriteFeature(FEAT_META_INFO, buf);
+  return WriteFeature(FEAT_META_INFO, buf.data(), buf.size());
 }
 
-bool RecordFileWriter::WriteFeature(int feature, const std::vector<char>& data) {
-  return WriteFeatureBegin(feature) && Write(data.data(), data.size()) && WriteFeatureEnd(feature);
+bool RecordFileWriter::WriteDebugUnwindFeature(const DebugUnwindFeature& debug_unwind) {
+  GOOGLE_PROTOBUF_VERIFY_VERSION;
+  proto::DebugUnwindFeature proto_debug_unwind;
+  for (auto& file : debug_unwind) {
+    auto proto_file = proto_debug_unwind.add_file();
+    proto_file->set_path(file.path);
+    proto_file->set_size(file.size);
+  }
+  std::string s;
+  if (!proto_debug_unwind.SerializeToString(&s)) {
+    LOG(ERROR) << "SerializeToString() failed";
+    return false;
+  }
+  return WriteFeature(FEAT_DEBUG_UNWIND, s.data(), s.size());
+}
+
+bool RecordFileWriter::WriteFeature(int feature, const char* data, size_t size) {
+  return WriteFeatureBegin(feature) && Write(data, size) && WriteFeatureEnd(feature);
 }
 
 bool RecordFileWriter::WriteFeatureBegin(int feature) {
@@ -491,10 +515,12 @@ bool RecordFileWriter::Close() {
     result = false;
   }
 
-  if (fclose(record_fp_) != 0) {
+  if (own_fp_ && fclose(record_fp_) != 0) {
     PLOG(ERROR) << "failed to close record file '" << filename_ << "'";
     result = false;
   }
   record_fp_ = nullptr;
   return result;
 }
+
+}  // namespace simpleperf
